Refactor and fix hdx_acap api
