+++ task_geo/data_sources/__init__.py
@@ -1,8 +1,8 @@
 from task_geo.data_sources.covid.cds import cds
 from task_geo.data_sources.covid.nyt import nyt
 from task_geo.data_sources.demographics.us_census import us_census
-from task_geo.data_sources.hdx_acap import hdx_acap
-from task_geo.data_sources.hdx_oxfd import hdx_oxfd
+from task_geo.data_sources.social_distancing.hdx_acap import hdx_acap
+from task_geo.data_sources.social_distancing.hdx_oxford import hdx_oxford
 from task_geo.data_sources.noaa import noaa_api
 
 AVAILABLE_DATA_SOURCES = {
@@ -11,7 +11,7 @@ AVAILABLE_DATA_SOURCES = {
     'us_census': us_census,
     'nyt': nyt,
     'hdx_acap': hdx_acap,
-    'hdx_oxfd': hdx_oxfd
+    'hdx_oxford': hdx_oxford
 }
