+++ task_geo/data_sources/social_distancing/hdx_acap/__init__.py
@@ -1,12 +1,9 @@
-from task_geo.data_sources.hdx_acap.hdx_acap_connector import hdx_acap_connector
-from task_geo.data_sources.hdx_acap.hdx_acap_formatter import hdx_acap_formatter
+from task_geo.data_sources.social_distancing.hdx_acap.hdx_acap_connector import hdx_acap_connector
+from task_geo.data_sources.social_distancing.hdx_acap.hdx_acap_formatter import hdx_acap_formatter
 
 
 def hdx_acap():
     """Retrieves formatted Government Measures Dataset from HDX DataSource
-
-    Arguments:
-        None
 
     Returns:
         pandas.DataFrame
@@ -19,3 +16,7 @@ def hdx_acap():
     """
     raw = hdx_acap_connector()
     return hdx_acap_formatter(raw)
+
+
+if __name__ == '__main__':
+    hdx_acap()
